support for firefox
Restructuring the code to include firefox & chrome extensions
