+++ web2py-devmode-chrome/contentscript.js
