+++ web2py-devmode-chrome/manifest.json
