+++ web2py-devmode-firefox/lib/main.js
@@ -1,0 +1,8 @@
+var data = require("sdk/self").data;
+var pageMod = require("sdk/page-mod");
+
+pageMod.PageMod({
+    //TODO: Require a regex to include only localhost, 127.0.0.1 pages
+	include : "*",
+	contentScriptFile : data.url("contentscript.js")
+});
